LayoutTests/imported/w3c/web-platform-tests/svg/animations/svgcolor-animation-invalid-value-1-expected.txt


FAIL
  Test SVGColor animation with invalid value: Invalid color string in from attribute
 assert_equals: expected "rgb(0, 128, 0)" but got "rgba(170, 0, 0, 0.667)"
